feat: separate store and update usage in TracksView
